pyproject.toml
[project]
name = "bachelor-thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "calplot>=0.1.7.5",
    "fastapi>=0.135.1",
    "garminconnect>=0.2.38",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
